// generated-sources/openapi/model/tagDto.ts
/**
 * Prestito
 * Prestito
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TagDto { 
    readonly id?: number;
    readonly created_date?: string;
    readonly modified_date?: string;
    title?: string;
}